refactor: update import paths to use absolute paths for consistency
